camera-stream/requirements.txt
flask==3.0.0
requests==2.31.0
urllib3==2.1.0
python-dotenv==1.0.0
Pillow==10.1.0
opencv-python-headless==4.8.1.78
ultralytics==8.3.0
numpy==1.24.3
torch==2.6.0
torchvision==0.21.0
